update heap while unwinding stack for pretty printing step
Latest reduction is held on Control while the stack has the Update continuation for the
thunks.

Printing occurs with Update still to be performed leading any other expression that
indirects to the same (now reduced) expression to still hold the old value, hence printing
it incorrectly.

While unwinding the stack, any Update continuation found updates a local copy of the heap.

Inelegant since this local copy is discarded, later when machine actually transitions the
update will be again performed.